+++ tcex/decorators/fail_on_output.py
@@ -75,7 +75,7 @@ class FailOnOutput:
                     raise RuntimeError(
                         'The fail_enabled value must be a boolean or resolved to bool.'
                     )
-            app.tcex.log.debug(f'Fail enabled is {enabled} ({self.fail_enabled}).')
+                app.tcex.log.debug(f'Fail enabled is {enabled} ({self.fail_enabled}).')
 
             failed = False
             if enabled: