+++ tcex/decorators/on_exception.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 """App Decorators Module."""
+import traceback
 import wrapt
 
 
@@ -65,12 +66,12 @@ class OnException:
                     raise RuntimeError(
                         'The exit_enabled value must be a boolean or resolved to bool.'
                     )
-            app.tcex.log.debug(f'Fail enabled is {enabled} ({self.exit_enabled}).')
+                app.tcex.log.debug(f'Fail enabled is {enabled} ({self.exit_enabled}).')
 
             try:
                 return wrapped(*args, **kwargs)
-            except Exception as e:
-                app.tcex.log.error(f'method failure ({e})')
+            except Exception:
+                app.tcex.log.error(traceback.format_exc())
                 app.exit_message = self.exit_msg  # for test cases
                 if enabled:
                     if self.write_output: