+++ tcex/testing/test_case_trigger_service.py
@@ -17,7 +17,8 @@ class TestCaseTriggerService(TestCaseServiceCommon):
         Returns:
             int: The App exit code
         """
-        args['tc_playbook_out_variables'] = ','.join(self.ij.output_variable_array)
+        # TODO: tc_playbook_out_variables should not be needed here as they are passed in config
+        # args['tc_playbook_out_variables'] = ','.join(self.ij.tc_playbook_out_variables)
         self.log.data('run', 'args', args)
         self.app = self.app_init(args)
 
@@ -37,13 +38,9 @@ class TestCaseTriggerService(TestCaseServiceCommon):
             self.app.tcex.service.heartbeat()
             self.app.tcex.service.ready = True
         except SystemExit as e:
-            if e.code != 0 and self.profile and e.code not in self.profile.exit_codes:
-                self.log.data(
-                    'run',
-                    'App failed',
-                    f'App exited with code of {e.code} in method run()',
-                    'error',
-                )
+            self.log.data(
+                'run', 'App failed', f'App exited with code of {e.code} in method run()', 'error',
+            )
             return self._exit(e.code)
         except Exception:
             self.log.data(